Implement stage strategy pattern for SQL handling
- Introduced StageStrategy abstract class to define a common interface for stage-specific logic.
- Created StageStrategyRegistry to manage and retrieve strategies for different stages.
- Developed individual strategies for each SQL-related stage:
  - AskSqlMethodStrategy: Handles user input for SQL method selection.
  - NeedNaturalLanguageStrategy: Generates SQL from natural language descriptions.
  - NeedUserSqlStrategy: Processes user-provided SQL queries.
  - ConfirmGeneratedSqlStrategy: Manages confirmation of generated SQL.
  - ConfirmUserSqlStrategy: Handles confirmation of user-provided SQL.
  - ExecuteSqlStrategy: Executes SQL jobs and manages parameter gathering.
  - ShowResultsStrategy: Displays results after SQL execution.
  - NeedWriteOrEmailStrategy: Manages post-execution options for results.
- Added logging for better traceability and debugging.
- Ensured strategies adhere to SOLID principles for maintainability and extensibility.
